feat: Support retrieving the head SHA from `workflow_run`
For secure workflow as Github officially advises, a `pull_request` event should perform a build for untrusted code and then make that available to a `workflow_run` triggered workflow when secrets are needed. Such as deployment. As the 2nd workflow (`workflow_run`) is run from the main/production branch (eg master), it will have the wrong `context.sha` value.

The head commit SHA is available matching that of the pull_request event, using this will allow the action to apply it's features correctly to the intended commit on pull request, not on production branch latest commit.