chore: Provide `workflow_run` troubleshooting
Adds a troubleshooting section for users to learn about `pull_request` workflow gotcha and an example of how to get it setup. Includes information on features that presently are unsupported for this action with `workflow_run` and a workaround for PR comment and status updates in the meantime. Deployment environment updates can't be disabled and will use the wrong commit for now.